Implement eth_getWorkWithAppendedExtraData
Previously, when any pool node queries with getWork(), all nodes would
receieve the same work. This exposes the pool to a vulnerability where
if there are multiple difficulty ports, a miner can just send their
share to the port with the highest valid difficulty to unfairly claim
increased rewards.

Implement this new API which will allow different ports to send out
different work that will not be valid on other ports.