Consolidate currentWork variables
Before, we were assuming that nil appendedExtraData would cause us to
get currentWork. However, currentWork is equal to currentWorks[""] so we
just consolidate these variables together.

In addition, we used to pass around string pointers to the
appendedExtraData but now we just consider nil and "" to be the same
case.